requirements.txt
azure-cognitiveservices-vision-customvision==3.1
setuptools_scm
azure.storage.blob
click
tqdm

black
pytest
pytest-env
flake8